fix(core): Resolve backend tests and cuda runtime linking issues - Fixed backends.go syntax error for go unit tests. - Fixed GPU state memory precision mismatch in CudaBackend mapping to cuDoubleComplex. - Fixed CMake CUDA target configuration for tests and native SM architecture targeting. - Added strict multi-qubit bounds validation to core gate methods.

+++ backend/src/Types.hpp
@@ -5,8 +5,8 @@ namespace qubit_engine {
 
 // --- Precision Settings ---
 // Uncomment the desired precision
-// using Precision = double;
-using Precision = float;
+using Precision = double;
+// using Precision = float;
 
 using Complex = std::complex<Precision>;
